GraphicsManager Development
Moved the map buffer to a StreamBuffer and added it to the graphics manager class.
All vbo, vao, and ibo objects are now created through the GraphicsManager.

+++ EmotionCore/src/Game/Text/TypewriterRichText.cs
@@ -148,8 +148,8 @@ namespace Emotion.Game.Text
             if (_characterEffectIndex == 0) return;
 
             // Draw the buffer.
-            if (!EffectFinished) _renderCache.SetRenderRange(0, _characterEffectIndex);
-            else _renderCache.SetRenderRange(0, _textStripped.Length);
+            if (!EffectFinished) _renderCache.SetRenderRange(0, (uint) _characterEffectIndex);
+            else _renderCache.SetRenderRange(0, (uint) _textStripped.Length);
 
             _renderCache.Render();
         }
